+++ NEWS.md
@@ -20,6 +20,7 @@ the dosing including dose amount and route.
   `PKNCAresults` objects.
 * `PKNCAdata()` is more restrictive on unknown arguments issuing an error
   when unknonwn arguments are present.
+* `intervals` argument to `PKNCAdata()` may now be a tibble (fixes #72).
 
 # PKNCA 0.8.5